Move Document from components directory to server/Document directory
